Add is-admin claim to JWTs & an endpoint to check for it
Notā bene:  the claim in the token isn't actually called  "is-admin",
but rather "tmcadm",  because the adminicity of a user is taken from
the TMC server:  TMC server admins are  also admins of  this program.
It is expected to be a JSON boolean value (true or false, not 0 or 1
or the strings "true" or "false"), and is technically optional:  the
JWT decoding routines default to false if it's not provided.