temake MasteryPercentage and add some cleaning up
Deleted old MasteryPercentage, it's controller and specs. Replaced with SkillMastery, a total new class with the same purpose. Made some specs and ran rubocop on model, controller and spec files.
Removed a couple of gems that were no longer in use.
Changed json-hash definitions from "errors" => to errors: so that they are consistent throughout the dashboard-api.
New file MockSkillMasteryData was made only to give us the skill-mastery-definitions that we are not getting anywhere else.